[ERROR] 에러 수정
