+++ final-pjt-front/paradiso_vue/src/App.vue
@@ -67,6 +67,7 @@ export default {
   },
   created: function() {
     const token = localStorage.getItem("accessToken")
+    const refreshToken = localStorage.getItem("refreshToken")
 
     if (token) {
       axios({
@@ -79,7 +80,20 @@ export default {
         })
         .catch((err) => {
           if (err.response.status === 401) {
-            this.$store.dispatch("refresh")
+            axios({
+              method: "post",
+              url: "http://127.0.0.1:8000/api/token/refresh/",
+              data: {'refresh': refreshToken},
+            })
+              .then((res) => {
+                localStorage.setItem("accessToken", res.data.access)
+                this.isLoggedIn = true;
+              })
+              .catch(() => {
+                localStorage.removeItem("accessToken")
+                localStorage.removeItem("refreshToken")
+                this.$router.push({ name: "LogInView" })
+              })
           }
         })
     }
