+++ final-pjt-front/paradiso_vue/src/store/index.js
@@ -1,7 +1,6 @@
 import Vue from 'vue'
 import Vuex from 'vuex'
 import createPersistedState from 'vuex-persistedstate'
-import axios from 'axios'
 
 Vue.use(Vuex)
 
@@ -30,25 +29,6 @@ export default new Vuex.Store({
 
     delete_user(context) {
       context.commit("DELETE_USER")
-    },
-
-    refresh() {
-      const refreshToken = localStorage.getItem("refreshToken")
-
-      axios({
-        method: "post",
-        url: "http://127.0.0.1:8000/api/token/refresh/",
-        data: {'refresh': refreshToken},
-      })
-        .then((res) => {
-          localStorage.setItem("accessToken", res.data.access)
-          this.isLoggedIn = true;
-        })
-        .catch(() => {
-          localStorage.removeItem("accessToken")
-          localStorage.removeItem("refreshToken")
-          this.$router.push({ name: "LogInView" })
-        })
     },
   },
   modules: {
